refactor(backend): use types crate instead of defining own types
