+++ types/src/lib.rs
@@ -1,1 +1,1 @@
-mod meetup;
+pub mod meetup;
