+++ types/src/meetup/search_response.rs
@@ -5,6 +5,6 @@ use super::search_request::{PageInfo, Result_};
 #[derive(Serialize, Deserialize, Debug, Default, PartialEq)]
 /// response object for /search
 pub struct Response {
-    page_info: PageInfo,
-    nodes: Vec<Result_>,
+    pub page_info: PageInfo,
+    pub nodes: Vec<Result_>,
 }
